requirements.txt
pydantic==2.9.2
requests==2.32.3
openai==1.52.2
python-dotenv==1.0.1
playwright==1.48.0
beautifulsoup4==4.12.3
